Add files via upload
This is an Uintah implementation of the Integrative model which is good for modeling B4C. This includes mechanisms of fracture + micromechanics + EoS + Amorphization + Granular flow (which also includes Cil's breakage model). The primary difference of this with PortableTongeRamesh model already implemented in Uintah 2.6 is Amorphization module. An Abaqus version of this implementation can also be obtained.